fix: dep cruiser and refactor
risk: low
