+++ libs/sdk-ui-dashboard/src/presentation/widget/insight/ViewModeDashboardInsight/InsightDrillDialog/useDrillDialogExportItems.ts
@@ -3,14 +3,13 @@
 import { IUiMenuInteractiveItem, IUiMenuItem } from "@gooddata/sdk-ui-kit";
 import { defineMessages, useIntl } from "react-intl";
 import {
-    IExecutionResultEnvelope,
     selectExecutionResultByRef,
     selectSettings,
     useDashboardSelector,
 } from "../../../../../model/index.js";
-import { isDataError, isDataErrorTooLarge } from "../../../../../_staging/errors/errorPredicates.js";
 import { idRef } from "@gooddata/sdk-model";
 import React from "react";
+import { getExportTooltip } from "../../../insightMenu/DefaultDashboardInsightMenu/getExportTooltips.js";
 
 export interface IMenuItemData {
     interactive: {
@@ -25,44 +24,11 @@ export type IMenuInteractiveItem = IUiMenuInteractiveItem<IMenuItemData>;
 
 export const DRILL_MODAL_EXECUTION_PSEUDO_REF = idRef("@@GDC_DRILL_MODAL");
 
-const tooltipMessages = defineMessages({
-    tooLarge: { id: "options.menu.data.too.large" },
-    rawError: { id: "options.menu.unsupported.raw.error" },
-    dataError: { id: "options.menu.unsupported.error" },
-    loading: { id: "options.menu.unsupported.loading" },
-    exporting: { id: "options.menu.export.in.progress" },
-});
-
 export const itemMessages = defineMessages({
     xlsx: { id: "widget.drill.dialog.exportToXLSX" },
     csvFormatted: { id: "widget.drill.dialog.exportToCSV.formatted" },
     csvRaw: { id: "widget.drill.dialog.exportToCSV.raw" },
 });
-
-export const getExportTooltip = ({
-    isRawExportsEnabled,
-    isExporting,
-    execution,
-}: {
-    execution?: IExecutionResultEnvelope;
-    isExporting: boolean;
-    isRawExportsEnabled?: boolean;
-}): string => {
-    if (isExporting) {
-        return tooltipMessages.exporting.id;
-    }
-
-    if (isDataErrorTooLarge(execution?.error)) {
-        return tooltipMessages.tooLarge.id;
-    } else if (isDataError(execution?.error)) {
-        if (isRawExportsEnabled) {
-            return tooltipMessages.rawError.id;
-        } else {
-            return tooltipMessages.dataError.id;
-        }
-    }
-    return tooltipMessages.loading.id;
-};
 
 export const useDrillDialogExportItems = ({
     isExporting,
